# doc/api/openapi.yaml
openapi: 3.0.0
info:
  title: FCA App Backend API
  version: 1.0.0
  description: Minimal API documentation for critical endpoints
servers:
  - url: http://localhost:3001
    description: Local Express Backend
paths:
...
  /api/companies/me:
    get:
      summary: Get current company details (Stubbed)
      responses:
        '401':
          description: Unauthorized (if no token)
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: string
        '200':
          description: Company details
          content:
            application/json:
              schema:
                type: object
                properties:
                  id:
                    type: string
                  name:
                    type: string
...